week03-homework-2/pyproject.toml
[project]
name = "week03-homework-2"
version = "0.1.0"
description = "RAG System with GraphRAG and Milvus"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "llama-index>=0.10.0",
    "llama-index-vector-stores-milvus>=0.1.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-llms-openai-like>=0.1.0",
    "pymilvus>=2.3.0",
    "neo4j>=5.17.0"
]

[tool.setuptools]
packages = ["milvus_faq", "graph_rag"]
